Create EventShiftDetails component, address width of autocomplete.
